+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## Unreleased
+
+### Changed
+
+- Dollar-number placeholders such as `$1` are no longer detected by default because they are indistinguishable
+  from currency values. Add `` String.raw`\$\d+` `` to `placeholderPatterns` to opt in.
+
 ## [0.7.2](https://github.com/xue-moe/i18n-smell-detector/compare/v0.7.1...v0.7.2) (2026-07-10)